feat(database): allow enforcing of SSL for database connection
Several ways devised to implement this easily. Most logical is to enforce this
feature only when actually running in production, but this does not work well
when using CI for testing. So a `TYPEORM_SSL_ENABLED` environment variable
provides more configuration flexibility (so you can also test this locally).

Originally, the path to the CA certificates file was hardcoded. However, this
does not allow for flexibility on non-standard platforms that store it in a
different location or specific requirements from the environment. As such, the
`TYPEORM_SSL_CACERTS` has been introduced to handle that.